fix: change url from :/publish/post/:postID to :/post/publish/:postID
